+++ Test/Unit/Model/QliroOrder/Builder/ValidateOrderBuilderStoreEmulationTest.php
@@ -218,6 +218,27 @@ class ValidateOrderBuilderStoreEmulationTest extends TestCase
                 return $this;
             }
 
+
+            public function getStreetLine($number)
+            {
+                return 'Sveavagen 1';
+            }
+
+            public function getCity(): string
+            {
+                return 'Stockholm';
+            }
+
+            public function getPostcode(): string
+            {
+                return '11122';
+            }
+
+            public function getCountryId(): string
+            {
+                return 'SE';
+            }
+
             public function getShippingInclTax(): float
             {
                 return 0.0;